[chore #160789425] : Refactor - minimize repetition by creating a helper_functions module
